+++ src/state/auto-scroller/fluid-scroller/get-scroll/adjust-for-size-limits.js
