+++ src/state/auto-scroller/fluid-scroller/get-scroll/get-on-axis/dampen-speed-by-duration.js
