+++ src/state/auto-scroller/fluid-scroller/get-scroll/get-on-axis/get-pixel-thresholds.js
